requirements.txt
pandas
numpy
matplotlib
seaborn
fastapi[standard]
finnhub-python
python-dotenv